requirements/base.txt
alabaster==0.7.10
argon2-cffi==18.1.0
attrs==17.4.0
awesome-slugify==1.6.5
Babel==2.5.3
backcall==0.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.6
django-filter==1.1.0
django-model-utils==3.1.1
django-redis==4.9.0
django-rest-auth==0.9.3
django-test-plus==1.0.22
djangorestframework==3.7.7
docutils==0.14
factory-boy==2.10.0
Faker==0.8.12
flake8==3.5.0
idna==2.6
imagesize==1.0.0
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
oauthlib==2.0.7
packaging==17.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
pkg-resources==0.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.5.0
pytest-django==3.1.2
pytest-sugar==0.9.1
python-dateutil==2.7.2
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
